Change the extension for all csaoid files from `.txt` to `.org`.
This allows my to include them in `org-brain`, thereby removing
duplication on my end. I can also remove transclusion in a few places
in the brain now, and directly link to these entries as children.

Since all the files were `.txt`, I wrote a small script to change the
extension. This is the command I ran:

    for f in *.txt; do echo "${f}"; done | awk -F '.' '{print $1}' - |
    xargs -I _ mv _.txt _.org

Brief explanation:
1. The `for` loop prints all the `.txt` files
2. The results are piped to `awk`, which treats `.` as the
   separator (instead of `<SPACE>` which is the default), and prints
   the first part (file-name)
3. This is piped to `xargs`, which takes the file-name and uses it in
   the placeholder in the command I've specified (`_` is the
   placeholder)

I am not sure if there is a clean way to do this in the do block
itself, but I'm happy with this short script as well.

These `.txt` files were already being read via `org-mode` because of
this line that I've added to the top of all files:

    -*- mode: org; comment-column: 0; -*-

which tells Emacs to load `org` as the major-mode.

Now that these files are `.org`, I don't need this line, so I removed
it as well using `sed`:

    for f in *.org; do sed -i '' '1{/^-\*- mode/d;}' ${f}; done;

I could have probably done this in a single command on Ubuntu, using
the `-s` operator in the `sed` command. I will try it out once I'm in
front of a Linux computer.